True HUD Implementation
In the real game the HUD is constantly checking if you are looking at a
valid wall to shoot a portal at. If you are, it displays both parts of the
HUD as filled in. if you arent they are both unfilled.

the hud in the game also has a portal indicator on the side of the hud
corresponding to the portal you last shot (also in the corresponding color)

both of the above things have been implemented in this PR. also attached are
images of it working.

I had to add a flag "just_checking" to a few functions involved when the
player normally shoots a portal and the validity checks are run. This flag
tells the functions to not actually shoot a new portal but just check if
shooting a portal would be valid.