Fixed Objects Clipping Through Walls
This fix was implemented by checking the distance from the player to
the nearest collision object (thats not the player and the object they're
holding) and if that is too close, adjusting the grab point back towards the player.

grab point is also bounded to -1.5 -> 0 so that the grab point doesnt go somewhere odd.

Also if a player is holding an object and they are much too close to a wall
(less than 0.3 distance) it will disable grabbing on the object and drop the
object.

Fixes #32
Fixes #11