Fixed Objects Dissapearing Between Rooms
previously if the player walked with a decore object to a new room
within a level, the object would dissapear (because it was no longer
set to the correct currentRoom). This can be seen in test chamber 0 with the
radio for example.

- this is a simple fix to change the currentRoom of a grabbed object
to the players room if the player walks through a doorway.