Correct Chamber Numbers
chamber numbers displayed in both the save and load screen menu's are now the correct number.
I acheived this by making a function that creates a mapping from level/room numbers -> chamber numbers

this mapping will need to be updated when new levels are added to the game, however its such low overhead that I think its okay. another option would be to store the chamber number when passing in or out of elevators and saving that number as the player progresses.

screenshots (of a couple chambers) attached.
all chambers verified to be working.