Sound Crackling Noise Fix
this bug was caused by too many sounds playing at once. to alleviate this two things were implemented

- lowered the max number of active sounds from 16 to 12, this based on my testing eliminated all crackling sounds
- made an array of sound clips that are "skippable". also added a function to check if a given sound ID is skippable based on the list.

the idea here is there are some sounds (footsteps, portal shooting, portal entering, etc.) that are acceptable to be skipped. there are other
sounds that are more important to not skip (glados speaking, menu sounds, buttons pressed, etc.). by marking some as skippable, those sounds
will not be able to be played after the number of active sounds passes a MAX_SKIPPABLE_SOUNDS threshold. this makes the event of skipping an
important sound like glados speaking MUCH less likely. as more sound effects are added to the game we will need to make sure to add them to
that list if they are less important.

Fixes #183