Fix Save Load Current Sound Problem
Fixes #382

- to alleviate the above issue I saved the current sound from each channel in the cutscene runner and serialize r/w appropriately.
- also made a new channel for Ambient Sounds called CH_AMBIENT. there are now 3 total channels CH_GLADOS, CH_MUSIC, and CH_AMBIENT

the only issue that is remaining is because we are relying on cutscene runner to que up the ambient sounds, the ambient sounds dont start until a cutscene starts on a level. this could be alleviated by making a trigger in the starting elevator of each level that simply triggers the ambient sound to play.