Allow changing base location of N64 toolchain
Environment variable N64_ROOT optionally sets the root directory of
the N64 toolchain.

To enable this to work with the linker script, the microcode object
files are now copied to the build directory before linking (since
there is no easy way to pass the path to the script dynamically).
An added benefit to this approach is that different microcode can
now be used easily if desired.