Automatic discovery of libgcc, remove newlib dependency (for now)
Implemented strcpy and strlen, which were all that were needed
(math functions and sprintf are in libultra).

To move further away from libultra, some standard library
implementation - such as newlib - will either need to be added
back later, or if the usage is small, functions can be reimplemented.
For now, removing it makes building easier.