Fix some build issues with Russian localization
The Russian localization does not translate everything. So when
building with Russian as the only language, some sound IDs will
be undefined. Check for this situation where relevant (turret).

Similarly, the sound ID generation logic expects the default
language to be a subset of the others. So if Russian is default
other languages will have sounds unnecessarily excluded (only
possible with Spanish due to sort order). This could be solved
by allowing arbitrary sounds in localized languages, but the
complexity is not needed since Russian is the only special
case. Instead, change the sort order to put Russian last
(localized sort order; same as for the text languages).