Decrease audio heap memory usage
Decrease max voices and sounds. The game's sound functions already
limit the number of active sounds to 12, so these don't need to be
this high.

This allows the heap size to be reduced back to 300 K.

More work will likely need to be done here as more localized sounds
are added.